refactoring to prepare for a second dataset
